requirements.txt
altgraph==0.17.4
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.43.0
cryptography==44.0.2
defusedxml==0.7.1
diff-match-patch==20241021
dj-database-url==2.3.0
Django==5.1.6
django-ckeditor==6.7.2
django-ckeditor-5==0.2.18
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-debug-toolbar==5.0.1
django-filter==25.1
django-import-export==4.3.7
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-mathfilters==1.0.0
django-taggit==6.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
drf-nested-routers==0.94.1
drf-yasg==1.21.10
Faker==36.1.1
idna==3.10
inflection==0.5.1
numpy==2.3.1
oauthlib==3.2.2
packaging==24.2
pandas==2.3.0
pefile==2023.2.7
pillow==11.1.0
psycopg-binary==3.2.4
psycopg2-binary==2.9.11
pycparser==2.22
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==80.9.0
shortuuid==1.0.13
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
sslcommerz-lib==1.0
tablib==3.8.0
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
